Optimize inference and training performance
Inference (VeinGrower): replace the iterative region-growing while loop
with a single batched pass over all image pixels. Previously a new
DataLoader was created at every growing iteration, causing repeated
Python/PyTorch overhead. Predictions are now accumulated in one pass and
the downstream thresholding/post-processing is unchanged.

Training (VeinGenerator): add n_samples parameter to cap the number of
foreground pixels sampled per image. Neighboring pixels produce near-
identical tiles (~98% overlap), so dense pixel sampling is redundant.
Random subsampling at 100K pixels/image reduces the dataset from ~9M to
~1.1M tiles per epoch while preserving spatial coverage; augmentation
(random rotations, flips, color jitter) maintains diversity across epochs.

Memory (ImageLoader): load images as float32 instead of float64, halving
CPU RAM usage with no impact on model precision since tiles are converted
to float32 before the GPU forward pass regardless.

Config: workers reduced from 32 to 4 to avoid per-worker RAM duplication
of the full image list.

+++ utils/ImageLoader.py
@@ -62,7 +62,7 @@ class ImageLoader:
     def load_image(self, path, pad=None):
 
         # load image as float array
-        image = np.array(Image.open(path), dtype=float)
+        image = np.array(Image.open(path), dtype=np.float32)
 
         # pre-process image
         if image.shape[2] > 3:
@@ -81,7 +81,7 @@ class ImageLoader:
             )
 
         # shape [H, W, 3]
-        return image.astype(float)
+        return image
 
     def load_mask(self, path, pad=None):
